p4/pyproject.toml
[project]
name = "fdi-pln-2602-p4"
version = "1.0.0"
description = "Sistema de recuperación de información sobre el Quijote para la práctica 4 de PLN"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "click>=8.3.1",
    "dynaconf>=3.2.0",
    "loguru>=0.7.3",
    "nltk>=3.9.0",
    "numpy>=2.0.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.1.0",
    "rich>=14.3.2",
    "spacy>=3.8.0",
    "textual>=0.85.0",
    "typer>=0.21.1",
]

[tool.uv]
package = false
